# api-design/autogenerated/openapi.yaml
openapi: 3.0.0
info:
  title: TodoDoo API
  version: 0.0.0
...
paths:
...
  /api/funs:
    get:
      operationId: Funs_list
      description: List all funs for the authenticated user
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fun'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
    post:
      operationId: Funs_create
      description: Create a new fun
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fun'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFunRequest'
  /api/funs/{id}:
    get:
      operationId: Funs_get
      description: Get a specific fun by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fun'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
    patch:
      operationId: Funs_update
      description: Update a fun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fun'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFunRequest'
    delete:
      operationId: Funs_delete
      description: Delete a fun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
  /api/funs/{id}/reorder:
    post:
      operationId: Funs_reorder
      description: Reorder a fun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fun'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderTodoRequest'
...
security:
  - BearerAuth: []
components:
  schemas:
...
    CreateFunRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The fun title
        position:
          allOf:
            - $ref: '#/components/schemas/TodoPosition'
          description: Where to insert the fun (defaults to top)
      description: Request body for creating a new fun
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      description: Error response
...
    Fun:
      type: object
      required:
        - id
        - title
        - user_id
        - order
        - date_created
        - date_updated
      properties:
        id:
          type: string
          description: Unique identifier for the fun
        title:
          type: string
          description: The fun title
        user_id:
          type: string
          description: ID of the user who owns this fun
        order:
          type: string
          description: Order/position of the fun in the list
        date_created:
          type: string
          format: date-time
          description: When the fun was created
        date_updated:
          type: string
          format: date-time
          description: When the fun was last updated
      description: A fun item
...
    PositionAfter:
      type: object
      required:
        - type
        - id
      properties:
        type:
          type: string
          enum:
            - after
        id:
          type: string
          description: ID of the todo to insert after
      description: Insert after a specific todo
    PositionBottom:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - bottom
      description: Insert at the bottom of the list
    PositionTop:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - top
      description: Insert at the top of the list
    ReorderTodoRequest:
      type: object
      required:
        - position
      properties:
        position:
          allOf:
            - $ref: '#/components/schemas/TodoPosition'
          description: New position for the todo
      description: Request body for reordering a todo
...
    TodoPosition:
      anyOf:
        - $ref: '#/components/schemas/PositionTop'
        - $ref: '#/components/schemas/PositionBottom'
        - $ref: '#/components/schemas/PositionAfter'
      description: Position for inserting a new todo
...
    UpdateFunRequest:
      type: object
      properties:
        title:
          type: string
          description: The fun title
      description: Request body for updating a fun
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://localhost:8000
    description: Development server
    variables: {}